+++ src-tauri/src/lib.rs
@@ -22,6 +22,7 @@ pub fn run() {
         open_xliff
     ]);
     tauri::Builder::default()
+        .plugin(tauri_plugin_updater::Builder::new().build())
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_store::Builder::new().build())
         .plugin(tauri_plugin_dialog::init())