+++ src/components/window/titlebar.tsx
@@ -1,27 +1,23 @@
-import { BookOpen, Code, Grid2X2Plus, Info, MessageCircleWarning, Settings } from "lucide-react"
-import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuLabel, DropdownMenuSeparator, DropdownMenuTrigger } from "@/components/ui/dropdown-menu";
-import { SiGithub } from "react-icons/si"
-import { openUrl } from "@tauri-apps/plugin-opener"
 import { Copy, Minus, Square, X } from "lucide-react"
 import { Button } from "@/components/ui/button"
 import { ButtonGroup } from "@/components/ui/button-group"
-import MenuBar from "./menubar";
 import { getCurrentWindow } from "@tauri-apps/api/window";
-import { useState, useEffect } from "react";
+import { useState, useEffect, lazy, Suspense } from "react";
 import { message } from "@tauri-apps/plugin-dialog";
 import { useAppTranslation } from "@/context/translation";
 import FileActions from "@/actions/file";
 import { toast } from "sonner";
 import { getErrorMessage } from "@/lib/utils";
-import AboutPopup from "@/popups/about";
-import SettingsPopup from "@/popups/settings";
+import { MenubarLoader } from "../../loaders/titlebar";
+
+const LogoDropdown = lazy(()=>import("./logo-dropdown"));
+const MenuBar = lazy(()=>import("./menubar"))
 
 interface TitleBarProps{
-     hideMenubar?: boolean,
      title?: string,
      hideMaximize?: boolean
 }
-export default function TitleBar({hideMaximize, hideMenubar, title}: TitleBarProps){
+export default function TitleBar({hideMaximize, title}: TitleBarProps){
      const appWindow = getCurrentWindow()
      const [isMaximized, setIsMaximized] = useState(false)
      const {isDirty, files, setIsDirty, table, langs} = useAppTranslation()
@@ -72,58 +68,12 @@ export default function TitleBar({hideMaximize, hideMenubar, title}: TitleBarPro
      }, [])
      return (
           <div className="flex items-center justify-between gap-2 bg-secondary/80 dark:bg-card/80 text-secondary-foreground border-b shadow-xs pl-2 sticky top-0 left-0 z-30 w-full h-9 backdrop-blur-md">
-               <div className="flex items-center gap-2">
-                    <DropdownMenu modal={false}>
-                         <DropdownMenuTrigger>
-                              <img src="/logo.png" alt="I18N Translator" width={24} height={24} className="select-none"/> 
-                         </DropdownMenuTrigger>
-                         <DropdownMenuContent className="w-full min-w-32">
-                              <DropdownMenuLabel>{title}</DropdownMenuLabel>
-                              {!hideMenubar && (
-                                   <>
-                                   <DropdownMenuSeparator/>
-                                   <AboutPopup triggerButton={(
-                                        <DropdownMenuItem onSelect={e=>e.preventDefault()}>
-                                             <Info className="text-muted-foreground"/>
-                                             About I18N Translator
-                                        </DropdownMenuItem>
-                                   )}/>
-                                   <SettingsPopup triggerButton={(
-                                        <DropdownMenuItem onSelect={e=>e.preventDefault()}>
-                                             <Settings className="text-muted-foreground"/>
-                                             Settings
-                                        </DropdownMenuItem>
-                                   )}/>
-                                   </>
-                              )}
-                              <DropdownMenuSeparator/>
-                              <DropdownMenuItem onClick={()=>openUrl("https://github.com/ArsenTech/i18n-translator")}>
-                                   <SiGithub className="text-muted-foreground"/>
-                                   Github
-                              </DropdownMenuItem>
-                              <DropdownMenuItem onClick={()=>openUrl("https://github.com/ArsenTech/i18n-translator/tree/main/docs")}>
-                                   <BookOpen className="text-muted-foreground"/>
-                                   Documentation
-                              </DropdownMenuItem>
-                              <DropdownMenuItem onClick={()=>openUrl("https://github.com/ArsenTech/i18n-translator/blob/main/docs/CONTRIBUTING.md")}>
-                                   <Code className="text-muted-foreground"/>
-                                   Contribute
-                              </DropdownMenuItem>
-                              <DropdownMenuSeparator/>
-                              <DropdownMenuItem onClick={()=>openUrl("https://github.com/ArsenTech/i18n-translator/issues/new?assignees=&labels=&template=bug_report.md&title=")}>
-                                   <MessageCircleWarning className="text-muted-foreground"/>
-                                   Report a bug
-                              </DropdownMenuItem>
-                              <DropdownMenuItem onClick={()=>openUrl("https://github.com/ArsenTech/i18n-translator/issues/new?assignees=&labels=&template=feature_request.md&title=")}>
-                                   <Grid2X2Plus className="text-muted-foreground"/>
-                                   Request a feature
-                              </DropdownMenuItem>
-                         </DropdownMenuContent>
-                    </DropdownMenu>
-                    {!hideMenubar && (
+               <Suspense fallback={<MenubarLoader/>}>
+                    <div className="flex items-center gap-2">
+                         <LogoDropdown title={title}/>
                          <MenuBar/>
-                    )}
-               </div>
+                    </div>
+               </Suspense>
                <div
                     data-tauri-drag-region
                     className="flex-1 h-full"