+++ src/lib/types/enums.ts
@@ -5,3 +5,12 @@ export enum TranslationFormat{
      Resx = "resx",
      Xliff = "xliff",
 }
+export enum UpdaterStatus {
+     Checking = "checking",
+     Updating = "updating",
+     NeedsUpdate = "needs-update",
+     Updated = "updated",
+     CheckError = "failed-check",
+     Completed = "completed",
+     UpdateError = "failed-update"
+}